tracer.extract() might return incomplete spancontext
Make sure that traceId and spanId is set in SpanContext from tracer.extract

BraveTracer will return an incomplete SpanContext where toTraceId()
and toSpanId() both return null instead of a null SpanContext if
the extract did not find any information in the carrier.

This commit verifiers that the returned SpanContext actually
has both trace and span id.